pcapimport + html_format + html_constructor
Combined Mike's pcapimport.py with Chris's html_format.py and html_constructor.py scripts.

Prompts user for input, verifies both the "high, medium, low" text files, as well as the ".pcap" are valid. If neither of these files are valid, the program will exit with an error, specifying what is missing and how to correct said error (full path/location needed, wrong directory, etc.) 

If the inputs are valid, the script will run as normal, and open a firefox web browser with the constructed map. If firefox is not installed, the "mapper.html" file can be opened in the same directory as the script upon successful completion, to view the map.